Moved tab order functions to ui/tools/urlutils.js file, updated components and test cases accordingly
