+++ src/core/core.js
@@ -9,8 +9,6 @@ import QuestionSubmission from './models/questionsubmission';
 import AnalyticsEvent from './analytics/analyticsevent';
 import FilterRegistry from './filters/filterregistry';
 
-import SearchParams from '../../src/ui/dom/searchparams';
-
 /** @typedef {import('./services/searchservice').default} SearchService */
 /** @typedef {import('./services/autocompleteservice').default} AutoCompleteService */
 /** @typedef {import('./services/questionanswerservice').default} QuestionAnswerService */
@@ -453,88 +451,5 @@ export default class Core {
 
   on (evt, moduleId, cb) {
     return this.globalStorage.on(evt, moduleId, cb);
-  }
-
-  getUrlParams () {
-    return new SearchParams(window.location.search.substring(1));
-  }
-
-  getDefaultTabOrder (tabsConfig, urlParams) {
-    let tabOrder = [];
-
-    // Use the ordering from the URL as the primary configuration
-    // And then merge it with the local configuration, if provided.
-    if (urlParams && urlParams.has('tabOrder')) {
-      tabOrder = urlParams.get('tabOrder').split(',');
-    }
-
-    for (let i = 0; i < tabsConfig.length; i++) {
-      const tab = tabsConfig[i];
-      // Some tabs don't have verticalKey, so we map it from URL
-      if (!tab.verticalKey) {
-        tab.verticalKey = tab.url;
-      }
-      // Avoid duplicates if config was provided from URL
-      if (tabOrder.includes(tab.verticalKey)) {
-        continue;
-      }
-
-      // isFirst should always be the first element in the list
-      if (tab.isFirst) {
-        tabOrder.unshift(tab.verticalKey);
-      } else {
-        tabOrder.push(tab.verticalKey);
-      }
-    }
-
-    return tabOrder;
-  }
-
-  /**
-   * mergeTabOrder merges two arrays into one
-   * by appending additional tabs to the end of the original array
-   * @param {string[]} tabOrder Tab order provided by the server
-   * @param {string[]} otherTabOrder Tab order provided by configuration
-   * @return {string[]}
-   */
-  mergeTabOrder (tabOrder, otherTabOrder, tabs) {
-    for (let i = 0; i < otherTabOrder.length; i++) {
-      const tabConfig = otherTabOrder[i];
-      if (tabOrder.includes(tabConfig)) {
-        continue;
-      }
-      // isFirst should be an override to dynamic tab ordering.
-      if (tabs[tabConfig] && tabs[tabConfig].isFirst) {
-        tabOrder.unshift(tabConfig);
-      } else {
-        tabOrder.push(tabConfig);
-      }
-    }
-
-    return tabOrder;
-  }
-
-  generateTabUrl (baseUrl, params, tabs, tabOrder) {
-    const context = this.globalStorage.getState(StorageKeys.API_CONTEXT);
-    if (context) {
-      params.set(StorageKeys.API_CONTEXT, context);
-    }
-    const referrerPageUrl = this.globalStorage.getState(StorageKeys.REFERRER_PAGE_URL);
-    if (referrerPageUrl !== null) {
-      params.set(StorageKeys.REFERRER_PAGE_URL, referrerPageUrl);
-    }
-
-    // We want to persist the params from the existing URL to the new
-    // URLS we create.
-    let dataTabOrder = [];
-    if (this.globalStorage.getState(StorageKeys.NAVIGATION)) {
-      dataTabOrder = this.globalStorage.getState(StorageKeys.NAVIGATION).tabOrder;
-    }
-    if (tabOrder !== undefined && dataTabOrder !== undefined) {
-      tabOrder = this.mergeTabOrder(dataTabOrder, tabOrder, tabs);
-      params.set('tabOrder', tabOrder);
-    }
-
-    return baseUrl + '?' + params.toString();
   }
 }
