+++ src/ui/components/navigation/navigationcomponent.js
@@ -6,6 +6,7 @@ import Component from '../component';
 import { AnswersComponentError } from '../../../core/errors/errors';
 import StorageKeys from '../../../core/storage/storagekeys';
 import DOM from '../../dom/dom';
+import { generateTabUrl, mergeTabOrder, getDefaultTabOrder, getUrlParams } from '../../tools/urlutils';
 
 /**
  * The debounce duration for resize events
@@ -156,7 +157,7 @@ export default class NavigationComponent extends Component {
      * @type {Array.<String>} The list of VS verticalKeys
      * @private
      */
-    this._tabOrder = this.core.getDefaultTabOrder(this._tabsConfig, this.core.getUrlParams());
+    this._tabOrder = getDefaultTabOrder(this._tabsConfig, getUrlParams());
 
     /**
      * Breakpoints at which navigation items move to the "more" dropdown
@@ -323,7 +324,7 @@ export default class NavigationComponent extends Component {
    */
   setState (data = {}) {
     if (data.tabOrder !== undefined) {
-      this._tabOrder = this.core.mergeTabOrder(data.tabOrder, this._tabOrder, this._tabs);
+      this._tabOrder = mergeTabOrder(data.tabOrder, this._tabOrder, this._tabs);
     }
 
     // Since the tab ordering can change based on the server data
@@ -333,7 +334,7 @@ export default class NavigationComponent extends Component {
     for (let i = 0; i < this._tabOrder.length; i++) {
       let tab = this._tabs[this._tabOrder[i]];
       if (tab !== undefined) {
-        tab.url = this.core.generateTabUrl(tab.baseUrl, this.core.getUrlParams(), this._tabs, this._tabOrder);
+        tab.url = generateTabUrl(tab.baseUrl, getUrlParams(), this._tabOrder);
         tabs.push(tab);
       }
     }
@@ -377,10 +378,6 @@ export default class NavigationComponent extends Component {
     return el.closest(closestElSelector);
   }
 
-  // getUrlParams () {
-  //   return new SearchParams(window.location.search.substring(1));
-  // }
-
   shouldCollapse () {
     switch (this._mobileOverflowBehavior) {
       case MOBILE_OVERFLOW_BEHAVIOR_OPTION.COLLAPSE:
@@ -391,75 +388,4 @@ export default class NavigationComponent extends Component {
         return navWidth > MOBILE_BREAKPOINT;
     }
   }
-
-  // /**
-  //  * getDefaultTabOrder will compute the initial tab ordering based
-  //  * on a combination of the configuration provided directly to the component
-  //  * and the url params.
-  //  * @param {Object[]} tabsConfig
-  //  * @param {SearchParams}
-  //  */
-  // getDefaultTabOrder (tabsConfig, urlParams) {
-  //   let tabOrder = [];
-
-  //   // Use the ordering from the URL as the primary configuration
-  //   // And then merge it with the local configuration, if provided.
-  //   if (urlParams && urlParams.has('tabOrder')) {
-  //     tabOrder = urlParams.get('tabOrder').split(',');
-  //   }
-
-  //   for (let i = 0; i < tabsConfig.length; i++) {
-  //     const tab = tabsConfig[i];
-  //     // Some tabs don't have verticalKey, so we map it from URL
-  //     if (!tab.verticalKey) {
-  //       tab.verticalKey = tab.url;
-  //     }
-
-  //     // Avoid duplicates if config was provided from URL
-  //     if (tabOrder.includes(tab.verticalKey)) {
-  //       continue;
-  //     }
-
-  //     // isFirst should always be the first element in the list
-  //     if (tab.isFirst) {
-  //       tabOrder.unshift(tab.verticalKey);
-  //     } else {
-  //       tabOrder.push(tab.verticalKey);
-  //     }
-  //   }
-
-  //   return tabOrder;
-  // }
-
-  // /**
-  //  * mergeTabOrder merges two arrays into one
-  //  * by appending additional tabs to the end of the original array
-  //  * @param {string[]} tabOrder Tab order provided by the server
-  //  * @param {string[]} otherTabOrder Tab order provided by configuration
-  //  * @return {string[]}
-  //  */
-  // mergeTabOrder (tabOrder, otherTabOrder) {
-  //   for (let i = 0; i < otherTabOrder.length; i++) {
-  //     const tabConfig = otherTabOrder[i];
-  //     if (tabOrder.includes(tabConfig)) {
-  //       continue;
-  //     }
-
-  //     // isFirst should be an override to dynamic tab ordering.
-  //     if (this._tabs[tabConfig] && this._tabs[tabConfig].isFirst) {
-  //       tabOrder.unshift(tabConfig);
-  //     } else {
-  //       tabOrder.push(tabConfig);
-  //     }
-  //   }
-
-  //   return tabOrder;
-  // }
-
-  // generateTabUrl (baseUrl, params = new URLSearchParams()) {
-  //   // We want to persist the params from the existing URL to the new
-  //   // URLS we create.
-  //   params.set('tabOrder', this._tabOrder);
-  //   return baseUrl + '?' + params.toString();
-  // }
 }
