+++ src/ui/sass/modules/_FilterOptions.scss
@@ -132,6 +132,9 @@
     align-items: center;
     margin-left: 22px;
     cursor: pointer;
+    word-break: break-word;
+    hyphens: auto;
+    width: 90%;
 
     /* Checkmark outer box */
     &::before {
